Standardized moderation effect given an lm output

Compute the standardized moderation effect given an lm output.

stdmod(
  lm_out,
  x = NULL,
  w = NULL,
  y = NULL,
  x_rescale = TRUE,
  w_rescale = TRUE,
  y_rescale = TRUE
)

stdmod_boot(
  lm_out,
  ...,
  nboot = 100,
  conf = 0.95,
  boot_args = NULL,
  full_output = FALSE
)

Arguments

lm_out

The output from lm.

x

The independent variable, that is, the variable with its effect being moderated. If supplied, it's standard deviation will be used for rescaling. Default is NULL.

w

The moderator. If supplied, it's standard deviation will be used for rescaling. Default is NULL.

y

The dependent (outcome) variable. If supplied, it's standard deviation will be used for rescaling. Default is NULL.

x_rescale

If TRUE, will rescale x by its SD. Default is TRUE.

w_rescale

If TRUE, will rescale w by its SD. Default is TRUE.

y_rescale

If TRUE, will rescale y by its SD. Default is TRUE.

...

Parameters to be passed to stdmod

nboot

The number of bootstrapping samples. Default is 100

conf

The level of confidence for the confidence interval. Default is .95.

boot_args

A named list of arguments to be passed to boot. Default is NULL.

full_output

Whether the full output from boot is return. Default is FALSE.

Value

The standardized moderation effect.

Details

Compute the standardized moderation effect given an lm output. Users specify the moderator, the predictor, the outcome variable, and the corresponding standardized moderation effect. Users can also select which variable(s) will be standardized.

stdmod_boot() is a wrapper of stdmod(). It computes the nonparametric bootstrapping confidence interval of the standardized moderation effect given the lm output.

Percentile interval from boot::boot.ci() is returned by this function. If other types of confidence intervals are desired, set full_output = TRUE and use boot::boot.ci() on the element boot_out in the output of this function.

Functions

stdmod: The base function compute standardized moderation effect

stdmod_boot: A wrapper of stdmod() that computes the nonparametric bootstrapping confidence interval of the standardized moderation effect

Examples

# Load a test data of 500 cases
# It has one predictor (iv), one moderator (mod), two covariates (v1 and v2),
# and one dv (dv). All variables continuous.
dat <- test_x_1_w_1_v_2_n_500

# Do regression as usual:
lm_raw <- lm(dv ~ iv*mod + v1 + v2, dat)
summary(lm_raw)
#> 
#> Call:
#> lm(formula = dv ~ iv * mod + v1 + v2, data = dat)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1650.87  -513.21     6.61   460.18  2189.74 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)  
#> (Intercept) 2814.565   4684.595   0.601   0.5482  
#> iv          -164.942    304.407  -0.542   0.5882  
#> mod          -17.223     46.420  -0.371   0.7108  
#> v1           -12.157     10.684  -1.138   0.2557  
#> v2            -4.284      6.290  -0.681   0.4962  
#> iv:mod         5.515      3.038   1.815   0.0701 .
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 708.1 on 494 degrees of freedom
#> Multiple R-squared:  0.5862,	Adjusted R-squared:  0.582 
#> F-statistic:   140 on 5 and 494 DF,  p-value: < 2.2e-16
#> 

# The standard deivations of iv, dv, and mod:
sds <- apply(dat, 2, sd)
sds
#>          dv          iv         mod          v1          v2 
#> 1095.222413    1.991458    5.004342    2.970378    5.064710 

# Compute the standardized moderation effect:
stdmod_xyw <- stdmod(lm_raw, x = iv, y = dv, w = mod)
stdmod_xyw
#>     iv:mod 
#> 0.05018003 
# By default, all three variables will be standardized.

# Check against self-computed standardized moderation effect:
coef(lm_raw)["iv:mod"] * sds["iv"] * sds["mod"] / sds["dv"]
#>     iv:mod 
#> 0.05018003 

# Standardize only the iv, i.e., do not standardized dv and the moderator:
stdmod_x <- stdmod(lm_raw, x = iv, y = dv, w = mod, 
                   x_rescale = TRUE,  y_rescale = FALSE, w_rescale = FALSE)
stdmod_x
#>   iv:mod 
#> 10.98212 
# Check against self-computed moderation effect with only iv standardized:
coef(lm_raw)["iv:mod"] * sds["iv"]
#>   iv:mod 
#> 10.98212 

# \donttest{

dat <- test_x_1_w_1_v_2_n_500
# Do regression as usual:
lm_raw <- lm(dv ~ iv*mod + v1 + v2, dat)

# Compute the standardized moderation effect, generates its confidence interval 
# by nonparametric bootstrappinng.
set.seed(85740917)
stdmod_xyw_boot <- stdmod_boot(lm_raw, x = iv, w = mod, y = dv, nboot = 500)

# Print the ci
stdmod_xyw_boot$ci
#> [1] -0.002731078  0.107099811

# Repeat the analysis but keep the results from boot:
set.seed(85740917)
stdmod_xyw_boot <- stdmod_boot(lm_raw, x = iv, w = mod, y = dv, 
                                 nboot = 500, full_output = TRUE)

# Print the 95% percential confidence interval
stdmod_xyw_boot$ci
#> [1] -0.002731078  0.107099811

# Compute the confidence interval with other settings using boot.ci.
# For example, request the 90% BCa confidence interval:
library(boot)
boot.ci(stdmod_xyw_boot$boot_out, conf = .90, type = "bca")
#> BOOTSTRAP CONFIDENCE INTERVAL CALCULATIONS
#> Based on 500 bootstrap replicates
#> 
#> CALL : 
#> boot.ci(boot.out = stdmod_xyw_boot$boot_out, conf = 0.9, type = "bca")
#> 
#> Intervals : 
#> Level       BCa          
#> 90%   ( 0.0056,  0.0999 )  
#> Calculations and Intervals on Original Scale
# }