Standardized moderation effect and bootstrapping CI in lavaan

Compute the standardized moderation effect and bootstrapping CI given a lavaan output

stdmod_lavaan(fit, x, y, w, x_w, boot_ci = FALSE, R = 100, conf = 0.95, ...)

Arguments

fit
The SEM output by lavaan::lavaan() or its siblings.

x
The name of the independent variable in the model.

y
The name of the outcome variable in the model.

w
The name of the moderator in the model.

x_w
The name of the product term (x * w) in the model.

boot_ci
Boolean. Whether nonparametric bootstrapping will be conducted. Default is FALSE.

R
The number of nonparametric bootstrapping samples. Default is 100. Set this to at least 2000 in actual use.

conf
The level of confidence. Default is .95.

...
Optional arguments to be passed to boot::boot().

Value

A list with these elements:

stdmod: The standardized moderation effect.

ci: The nonparametric bootstrapping confidence interval. NA if confidence interval not requested.

boot_out: The raw output from boot::boot(). NA if confidence interval not requested.

Details

stdmod_lavaan() accepts a lavaan::lavaan object, the SEM output generated by lavaan::lavaan() and its sibling (e.g, lavaan::sem()) and compute the standardized moderation effect.

The standard deviations of the independent variable, moderator, and outcome are computed from the implied covariance matrix returned by lavaan::lavInspect().

If nonparametric bootstrapping confidence is requested with R bootstrapping samples, the model will be fitted R times to these samples, and the standardized moderation effect will be computed in each sample. This enasures that all componenets used in the computation, including the standard deviations, are also computed from the bootstrapping samples.

Note that the compuation can be slow because lavaan::lavaan() will be called R times.

Examples

# Load a test data of 500 cases
# It has one predictor (iv), one moderator (mod), two covariates (v1 and v2),
# and one dv (dv). All variables continuous.
dat <- test_x_1_w_1_v_2_n_500
library(lavaan)

#> This is lavaan 0.6-7

#> lavaan is BETA software! Please report any bugs.

# Scale down some variables to facilitate optimization in lavaan.
dat$dv <- (dat$dv - 6500)/1100
dat$mod <- (dat$mod - 95)/5
dat$iv <- (dat$iv - 10)
# Form the product term for the moderation
dat$iv_mod <- dat$iv * dat$mod

mod <- 
"
dv ~ iv + mod + mod + v1 + v2 + iv_mod
"
fit <- sem(mod, dat)
coef(fit)

#>     dv~iv    dv~mod     dv~v1     dv~v2 dv~iv_mod    dv~~dv 
#>     0.326     0.172    -0.011    -0.004     0.025     0.409 

# Compute the standardized moderation effect
out_noboot <- stdmod_lavaan(fit = fit, x = "iv",
                                y = "dv",
                                w = "mod",
                                x_w = "iv_mod")
out_noboot$stdmod

#>  dv~iv_mod 
#> 0.05012983 

# \donttest{
# Compute the standardized moderation effect and 
# its confidence interval based on nonparametric bootstrapping
set.seed(8479075)
system.time(out_boot <- stdmod_lavaan(fit = fit, x = "iv",
                                y = "dv",
                                w = "mod",
                                x_w = "iv_mod",
                          boot_ci = TRUE, R = 200))

#>    user  system elapsed 
#>    9.67    0.23   10.54 

# In real analysis, R should be at least 2000.
out_boot$ci

#>  lower_bound  upper_bound 
#> -0.002431213  0.103367432 

# Use boot.ci to compute the confidence interval using other settings
library(boot)
boot.ci(out_boot$boot_out, conf = .90, type = "perc")

#> BOOTSTRAP CONFIDENCE INTERVAL CALCULATIONS
#> Based on 200 bootstrap replicates
#> 
#> CALL : 
#> boot.ci(boot.out = out_boot$boot_out, conf = 0.9, type = "perc")
#> 
#> Intervals : 
#> Level     Percentile     
#> 90%   ( 0.0028,  0.0948 )  
#> Calculations and Intervals on Original Scale

# }