Standardized Moderation Effect and Bootstrap CI in 'lavaan'

Compute the standardized moderation effect in a structural equation model fitted by lavaan::lavaan() and form its nonparametric bootstrap confidence interval.

stdmod_lavaan(fit, x, y, w, x_w, boot_ci = FALSE, R = 100, conf = 0.95, ...)

Arguments

fit

The SEM output by lavaan::lavaan() or its siblings.

x

The name of the independent variable in the model.

y

The name of the outcome variable in the model.

w

The name of the moderator in the model.

x_w

The name of the product term (x * w) in the model.

boot_ci

Boolean. Whether nonparametric bootstrapping will be conducted. Default is FALSE.

R

The number of nonparametric bootstrapping samples. Default is 100. Set this to at least 2000 in actual use.

conf

The level of confidence. Default is .95.

...

Optional arguments to be passed to boot::boot().

Value

A list with these elements:

stdmod: The standardized moderation effect.

ci: The nonparametric bootstrap confidence interval. NA if confidence interval not requested.

boot_out: The raw output from boot::boot(). NA if confidence interval not requested.

Details

stdmod_lavaan() accepts a lavaan::lavaan object, the structural equation model output returned by lavaan::lavaan() and its wrappers (e.g, lavaan::sem()) and computes the standardized moderation effect using the formula in the appendix of Cheung, Cheung, Lau, Hui, and Vong (2022).

The standard deviations of the independent variable, moderator, and outcome variable are computed from the implied covariance matrix returned by lavaan::lavInspect(). Therefore, models fitted to data sets with missing data (e.g., with missing = "fiml") are also supported.

If nonparametric bootstrap confidence interval is requested with R bootstrap samples, the model will be fitted R times to these samples, and the standardized moderation effect will be computed in each sample. This ensures that all components used in the computation, including the standard deviations, are also computed from the bootstrapping samples.

Note that the computation can be slow because lavaan::lavaan() will be called R times.

References

Cheung, S. F., Cheung, S.-H., Lau, E. Y. Y., Hui, C. H., & Vong, W. N. (2022) Improving an old way to measure moderation effect in standardized units. Advance online publication. Health Psychology. doi:10.1037/hea0001188

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

# Load a test data of 500 cases
# It has one predictor (iv), one moderator (mod), two covariates (v1 and v2),
# and one dv (dv). All variables continuous.
dat <- test_x_1_w_1_v_2_n_500
library(lavaan)
#> This is lavaan 0.6-11
#> lavaan is FREE software! Please report any bugs.
# Scale down some variables to facilitate optimization in lavaan.
dat$dv <- (dat$dv - 6500)/1100
dat$mod <- (dat$mod - 95)/5
dat$iv <- (dat$iv - 10)
# Form the product term for the moderation
dat$iv_mod <- dat$iv * dat$mod

mod <-
"
dv ~ iv + mod + mod + v1 + v2 + iv_mod
"
fit <- sem(mod, dat)
coef(fit)
#>     dv~iv    dv~mod     dv~v1     dv~v2 dv~iv_mod    dv~~dv 
#>     0.326     0.172    -0.011    -0.004     0.025     0.409 

# Compute the standardized moderation effect
out_noboot <- stdmod_lavaan(fit = fit, x = "iv",
                                y = "dv",
                                w = "mod",
                                x_w = "iv_mod")
out_noboot$stdmod
#>  dv~iv_mod 
#> 0.05012982 

# Compute the standardized moderation effect and
# its confidence interval based on nonparametric bootstrapping
set.seed(8479075)
system.time(out_boot <- stdmod_lavaan(fit = fit, x = "iv",
                                y = "dv",
                                w = "mod",
                                x_w = "iv_mod",
                          boot_ci = TRUE, R = 50))
#>    user  system elapsed 
#>   1.056   0.008   1.065 
# In real analysis, R should be at least 2000.
out_boot$ci
#>  lower_bound  upper_bound 
#> -0.006182574  0.113850037 

# Use boot.ci to compute the confidence interval using other settings
library(boot)
boot.ci(out_boot$boot_out, conf = .90, type = "perc")
#> BOOTSTRAP CONFIDENCE INTERVAL CALCULATIONS
#> Based on 50 bootstrap replicates
#> 
#> CALL : 
#> boot.ci(boot.out = out_boot$boot_out, conf = 0.9, type = "perc")
#> 
#> Intervals : 
#> Level     Percentile     
#> 90%   ( 0.0156,  0.0971 )  
#> Calculations and Intervals on Original Scale
#> Some percentile intervals may be unstable